feat(obstacle-spawner): implement obstacle spawner UI and functionality
- Added ObstacleList component to display existing obstacles with actions to toggle, nudge, and remove them.
- Created ObstacleSpawnerForm for spawning new obstacles with fields for ID, type, position, dimensions, and movement patterns.
- Implemented buildObstacle utility to construct obstacle entities based on form input.
- Introduced reusable NumberField and CheckboxField components for form inputs.
- Developed formUtils for handling number conversion and unique ID generation.
- Styled components with CSS for a cohesive UI experience.
- Added types for obstacle entities and form state management.
- Implemented inspector UI for editing obstacle properties, including durability, flags, hazards, and movement patterns.
