+++ src/components/debug/obstacleSpawner/obstacleSpawner.css
@@ -1,0 +1,115 @@
+.obs-spawner-panel {
+  background: rgba(8, 10, 22, 0.82);
+  border: 1px solid rgba(255, 255, 255, 0.08);
+  color: #e9ecff;
+  padding: 12px;
+  border-radius: 10px;
+  max-width: 360px;
+  box-shadow: 0 12px 28px rgba(0, 0, 0, 0.35);
+  -webkit-backdrop-filter: blur(6px);
+  backdrop-filter: blur(6px);
+}
+
+.obs-spawner-title {
+  font-weight: 700;
+  margin-bottom: 8px;
+  letter-spacing: 0.2px;
+}
+
+.obs-spawner-grid-2 {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 8px;
+}
+
+.obs-spawner-mb-8 {
+  margin-bottom: 8px;
+}
+
+.obs-spawner-mb-12 {
+  margin-bottom: 12px;
+}
+
+.obs-spawner-stack {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.obs-spawner-label {
+  font-size: 12px;
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.obs-spawner-checkbox {
+  font-size: 12px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 8px;
+}
+
+.obs-spawner-input {
+  padding: 6px 8px;
+  border-radius: 6px;
+  border: 1px solid rgba(255, 255, 255, 0.08);
+  background: rgba(255, 255, 255, 0.06);
+  color: #f8f9ff;
+  font-size: 12px;
+}
+
+.obs-spawner-button {
+  padding: 10px 12px;
+  border-radius: 8px;
+  border: none;
+  cursor: pointer;
+  background: linear-gradient(135deg, #4c7bff, #7cd2ff);
+  color: #0b1024;
+  font-weight: 700;
+}
+
+.obs-spawner-button--full {
+  width: 100%;
+}
+
+.obs-spawner-list-title {
+  font-size: 12px;
+  opacity: 0.8;
+  margin-bottom: 8px;
+}
+
+.obs-spawner-list {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  max-height: 160px;
+  overflow-y: auto;
+}
+
+.obs-spawner-row {
+  display: grid;
+  grid-template-columns: 1fr auto;
+  gap: 6px;
+  padding: 6px 8px;
+  border-radius: 8px;
+  background: rgba(255, 255, 255, 0.04);
+  align-items: center;
+}
+
+.obs-spawner-row-title {
+  font-weight: 600;
+  font-size: 12px;
+}
+
+.obs-spawner-row-meta {
+  font-size: 11px;
+  opacity: 0.8;
+}
+
+.obs-spawner-row-actions {
+  display: flex;
+  gap: 4px;
+  align-items: center;
+}
